Move 'add_files' functionality into SDist classes.
